+++ phpunit.xml.dist
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <phpunit xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:noNamespaceSchemaLocation="https://schema.phpunit.de/10.5/phpunit.xsd" backupGlobals="false" colors="true"
-         bootstrap="tests/bootstrap.php" failOnRisky="false" failOnWarning="false">
+         bootstrap="tests/bootstrap.php" failOnRisky="true" failOnWarning="true">
     <php>
         <ini name="error_reporting" value="-1"/>
     </php>
