+++ tests/Application/Contracts/Bitrix24Accounts/Repository/Bitrix24AccountRepositoryInterfaceTest.php
@@ -2,17 +2,16 @@
 
 declare(strict_types=1);
 
-namespace Bitrix24\SDK\Tests\Unit\Application\Contracts\Bitrix24Accounts\Entity;
+namespace Bitrix24\SDK\Tests\Application\Contracts\Bitrix24Accounts\Repository;
 
-use Bitrix24\SDK\Application\ApplicationStatus;
 use Bitrix24\SDK\Application\Contracts\Bitrix24Accounts\Entity\Bitrix24AccountInterface;
 use Bitrix24\SDK\Application\Contracts\Bitrix24Accounts\Entity\Bitrix24AccountStatus;
+use Bitrix24\SDK\Application\Contracts\Bitrix24Accounts\Exceptions\Bitrix24AccountNotFoundException;
+use Bitrix24\SDK\Application\Contracts\Bitrix24Accounts\Repository\Bitrix24AccountRepositoryInterface;
 use Bitrix24\SDK\Core\Credentials\AuthToken;
 use Bitrix24\SDK\Core\Credentials\Scope;
 use Bitrix24\SDK\Core\Exceptions\InvalidArgumentException;
 use Bitrix24\SDK\Core\Exceptions\UnknownScopeCodeException;
-use Bitrix24\SDK\Core\Response\DTO\RenewedAuthToken;
-use Bitrix24\SDK\Services\Telephony\Call\Service\Call;
 use Carbon\CarbonImmutable;
 use Generator;
 use PHPUnit\Framework\Attributes\CoversClass;
@@ -21,9 +20,9 @@ use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\Attributes\TestDox;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Uid\Uuid;
-use Throwable;
+
 #[CoversClass(Bitrix24AccountInterface::class)]
-abstract class Bitrix24AccountInterfaceTest extends TestCase
+abstract class Bitrix24AccountRepositoryInterfaceTest extends TestCase
 {
     abstract protected function createBitrix24AccountImplementation(
         Uuid                  $uuid,
@@ -39,10 +38,15 @@ abstract class Bitrix24AccountInterfaceTest extends TestCase
         Scope                 $applicationScope
     ): Bitrix24AccountInterface;
 
+    abstract protected function createBitrix24AccountRepositoryImplementation(): Bitrix24AccountRepositoryInterface;
+
+    /**
+     * @throws Bitrix24AccountNotFoundException
+     */
     #[Test]
-    #[DataProvider('bitrix24AccountDataProvider')]
-    #[TestDox('test getId method')]
-    final public function testGetId(
+    #[DataProvider('bitrix24AccountForInstallDataProvider')]
+    #[TestDox('test save method for install start use case')]
+    final public function testSave(
         Uuid                  $uuid,
         int                   $bitrix24UserId,
         bool                  $isBitrix24UserAdmin,
@@ -57,243 +61,238 @@ abstract class Bitrix24AccountInterfaceTest extends TestCase
     ): void
     {
         $bitrix24Account = $this->createBitrix24AccountImplementation($uuid, $bitrix24UserId, $isBitrix24UserAdmin, $memberId, $domainUrl, $bitrix24AccountStatus, $authToken, $createdAt, $updatedAt, $applicationVersion, $applicationScope);
-        $this->assertEquals($uuid, $bitrix24Account->getId());
+
+        $bitrix24AccountRepository = $this->createBitrix24AccountRepositoryImplementation();
+
+        $bitrix24AccountRepository->save($bitrix24Account);
+
+        $acc = $bitrix24AccountRepository->getById($bitrix24Account->getId());
+        $this->assertEquals($bitrix24Account, $acc);
     }
 
     #[Test]
-    #[DataProvider('bitrix24AccountDataProvider')]
-    #[TestDox('test getBitrix24UserId method')]
-    final public function testGetBitrix24UserId(
-        Uuid                  $uuid,
-        int                   $bitrix24UserId,
-        bool                  $isBitrix24UserAdmin,
-        string                $memberId,
-        string                $domainUrl,
-        Bitrix24AccountStatus $bitrix24AccountStatus,
-        AuthToken             $authToken,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
-        int                   $applicationVersion,
-        Scope                 $applicationScope
-    ): void
+    #[TestDox('test getById method with non existing account')]
+    public function testGetByIdNotExists(): void
     {
-        $bitrix24Account = $this->createBitrix24AccountImplementation($uuid, $bitrix24UserId, $isBitrix24UserAdmin, $memberId, $domainUrl, $bitrix24AccountStatus, $authToken, $createdAt, $updatedAt, $applicationVersion, $applicationScope);
-        $this->assertEquals($bitrix24UserId, $bitrix24Account->getBitrix24UserId());
-    }
-
-    #[Test]
-    #[DataProvider('bitrix24AccountDataProvider')]
-    #[TestDox('test isBitrix24UserAdmin method')]
-    final public function testisBitrix24UserAdmin(
-        Uuid                  $uuid,
-        int                   $bitrix24UserId,
-        bool                  $isBitrix24UserAdmin,
-        string                $memberId,
-        string                $domainUrl,
-        Bitrix24AccountStatus $bitrix24AccountStatus,
-        AuthToken             $authToken,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
-        int                   $applicationVersion,
-        Scope                 $applicationScope
-    ): void
-    {
-        $bitrix24Account = $this->createBitrix24AccountImplementation($uuid, $bitrix24UserId, $isBitrix24UserAdmin, $memberId, $domainUrl, $bitrix24AccountStatus, $authToken, $createdAt, $updatedAt, $applicationVersion, $applicationScope);
-        $this->assertEquals($isBitrix24UserAdmin, $bitrix24Account->isBitrix24UserAdmin());
-    }
-
-    #[Test]
-    #[DataProvider('bitrix24AccountDataProvider')]
-    #[TestDox('test getMemberId method')]
-    final public function testGetMemberId(
-        Uuid                  $uuid,
-        int                   $bitrix24UserId,
-        bool                  $isBitrix24UserAdmin,
-        string                $memberId,
-        string                $domainUrl,
-        Bitrix24AccountStatus $bitrix24AccountStatus,
-        AuthToken             $authToken,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
-        int                   $applicationVersion,
-        Scope                 $applicationScope
-    ): void
-    {
-        $bitrix24Account = $this->createBitrix24AccountImplementation($uuid, $bitrix24UserId, $isBitrix24UserAdmin, $memberId, $domainUrl, $bitrix24AccountStatus, $authToken, $createdAt, $updatedAt, $applicationVersion, $applicationScope);
-        $this->assertEquals($memberId, $bitrix24Account->getMemberId());
-    }
-
-    #[Test]
-    #[DataProvider('bitrix24AccountDataProvider')]
-    #[TestDox('test getDomainUrl method')]
-    final public function testGetDomainUrl(
-        Uuid                  $uuid,
-        int                   $bitrix24UserId,
-        bool                  $isBitrix24UserAdmin,
-        string                $memberId,
-        string                $domainUrl,
-        Bitrix24AccountStatus $bitrix24AccountStatus,
-        AuthToken             $authToken,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
-        int                   $applicationVersion,
-        Scope                 $applicationScope
-    ): void
-    {
-        $bitrix24Account = $this->createBitrix24AccountImplementation($uuid, $bitrix24UserId, $isBitrix24UserAdmin, $memberId, $domainUrl, $bitrix24AccountStatus, $authToken, $createdAt, $updatedAt, $applicationVersion, $applicationScope);
-        $this->assertEquals($domainUrl, $bitrix24Account->getDomainUrl());
-    }
-
-    #[Test]
-    #[DataProvider('bitrix24AccountDataProvider')]
-    #[TestDox('test getStatus method')]
-    final public function testGetStatus(
-        Uuid                  $uuid,
-        int                   $bitrix24UserId,
-        bool                  $isBitrix24UserAdmin,
-        string                $memberId,
-        string                $domainUrl,
-        Bitrix24AccountStatus $bitrix24AccountStatus,
-        AuthToken             $authToken,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
-        int                   $applicationVersion,
-        Scope                 $applicationScope
-    ): void
-    {
-        $bitrix24Account = $this->createBitrix24AccountImplementation($uuid, $bitrix24UserId, $isBitrix24UserAdmin, $memberId, $domainUrl, $bitrix24AccountStatus, $authToken, $createdAt, $updatedAt, $applicationVersion, $applicationScope);
-        $this->assertEquals($bitrix24AccountStatus, $bitrix24Account->getStatus());
-    }
-
-    #[Test]
-    #[DataProvider('bitrix24AccountDataProvider')]
-    #[TestDox('test getAuthToken method')]
-    final public function testGetAuthToken(
-        Uuid                  $uuid,
-        int                   $bitrix24UserId,
-        bool                  $isBitrix24UserAdmin,
-        string                $memberId,
-        string                $domainUrl,
-        Bitrix24AccountStatus $bitrix24AccountStatus,
-        AuthToken             $authToken,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
-        int                   $applicationVersion,
-        Scope                 $applicationScope
-    ): void
-    {
-        $bitrix24Account = $this->createBitrix24AccountImplementation($uuid, $bitrix24UserId, $isBitrix24UserAdmin, $memberId, $domainUrl, $bitrix24AccountStatus, $authToken, $createdAt, $updatedAt, $applicationVersion, $applicationScope);
-        $this->assertEquals($authToken, $bitrix24Account->getAuthToken());
-    }
-
-    #[Test]
-    #[DataProvider('bitrix24AccountDataProvider')]
-    #[TestDox('test renewAuthToken method')]
-    final public function testRenewAuthToken(
-        Uuid                  $uuid,
-        int                   $bitrix24UserId,
-        bool                  $isBitrix24UserAdmin,
-        string                $memberId,
-        string                $domainUrl,
-        Bitrix24AccountStatus $bitrix24AccountStatus,
-        AuthToken             $authToken,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
-        int                   $applicationVersion,
-        Scope                 $applicationScope
-    ): void
-    {
-        $bitrix24Account = $this->createBitrix24AccountImplementation($uuid, $bitrix24UserId, $isBitrix24UserAdmin, $memberId, $domainUrl, $bitrix24AccountStatus, $authToken, $createdAt, $updatedAt, $applicationVersion, $applicationScope);
-        $newAuthToken = new AuthToken('access_token-2', 'refresh_token=2', 1609459202);
-        $applicationStatus = ApplicationStatus::subscription();
-
-        $renewedAuthToken = new RenewedAuthToken(
-            $newAuthToken,
-            $memberId,
-            'https://bitrix24.com/client',
-            'https://bitrix24.com/server',
-            $applicationStatus,
-            $domainUrl
-        );
-        $bitrix24Account->renewAuthToken($renewedAuthToken);
-
-
-        $this->assertEquals($newAuthToken, $bitrix24Account->getAuthToken());
-    }
-
-    #[Test]
-    #[DataProvider('bitrix24AccountDataProvider')]
-    #[TestDox('test getApplicationVersion method')]
-    final public function testGetApplicationVersion(
-        Uuid                  $uuid,
-        int                   $bitrix24UserId,
-        bool                  $isBitrix24UserAdmin,
-        string                $memberId,
-        string                $domainUrl,
-        Bitrix24AccountStatus $bitrix24AccountStatus,
-        AuthToken             $authToken,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
-        int                   $applicationVersion,
-        Scope                 $applicationScope
-    ): void
-    {
-        $bitrix24Account = $this->createBitrix24AccountImplementation($uuid, $bitrix24UserId, $isBitrix24UserAdmin, $memberId, $domainUrl, $bitrix24AccountStatus, $authToken, $createdAt, $updatedAt, $applicationVersion, $applicationScope);
-        $this->assertEquals($applicationVersion, $bitrix24Account->getApplicationVersion());
-    }
-
-
-    #[Test]
-    #[DataProvider('bitrix24AccountDataProvider')]
-    #[TestDox('test getApplicationScope method')]
-    final public function testGetApplicationScope(
-        Uuid                  $uuid,
-        int                   $bitrix24UserId,
-        bool                  $isBitrix24UserAdmin,
-        string                $memberId,
-        string                $domainUrl,
-        Bitrix24AccountStatus $bitrix24AccountStatus,
-        AuthToken             $authToken,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
-        int                   $applicationVersion,
-        Scope                 $applicationScope
-    ): void
-    {
-        $bitrix24Account = $this->createBitrix24AccountImplementation($uuid, $bitrix24UserId, $isBitrix24UserAdmin, $memberId, $domainUrl, $bitrix24AccountStatus, $authToken, $createdAt, $updatedAt, $applicationVersion, $applicationScope);
-        $this->assertEquals($applicationScope, $bitrix24Account->getApplicationScope());
-    }
-
-    #[Test]
-    #[DataProvider('bitrix24AccountDataProvider')]
-    #[TestDox('test changeDomainUrl method')]
-    final public function testChangeDomainUrl(
-        Uuid                  $uuid,
-        int                   $bitrix24UserId,
-        bool                  $isBitrix24UserAdmin,
-        string                $memberId,
-        string                $domainUrl,
-        Bitrix24AccountStatus $bitrix24AccountStatus,
-        AuthToken             $authToken,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
-        int                   $applicationVersion,
-        Scope                 $applicationScope
-    ): void
-    {
-        $newDomainUrl = 'new-bitrix24.com';
-        $bitrix24Account = $this->createBitrix24AccountImplementation($uuid, $bitrix24UserId, $isBitrix24UserAdmin, $memberId, $domainUrl, $bitrix24AccountStatus, $authToken, $createdAt, $updatedAt, $applicationVersion, $applicationScope);
-        $bitrix24Account->changeDomainUrl($newDomainUrl);
-        $this->assertEquals($newDomainUrl, $bitrix24Account->getDomainUrl());
+        $this->expectException(Bitrix24AccountNotFoundException::class);
+        $bitrix24AccountRepository = $this->createBitrix24AccountRepositoryImplementation();
+        $bitrix24AccountRepository->getById(Uuid::v7());
     }
 
     /**
+     * @throws InvalidArgumentException
+     * @throws Bitrix24AccountNotFoundException
+     */
+    #[Test]
+    #[DataProvider('bitrix24AccountForInstallDataProvider')]
+    #[TestDox('test delete method for happy path')]
+    final public function testDeleteHappyPath(
+        Uuid                  $uuid,
+        int                   $bitrix24UserId,
+        bool                  $isBitrix24UserAdmin,
+        string                $memberId,
+        string                $domainUrl,
+        Bitrix24AccountStatus $bitrix24AccountStatus,
+        AuthToken             $authToken,
+        CarbonImmutable       $createdAt,
+        CarbonImmutable       $updatedAt,
+        int                   $applicationVersion,
+        Scope                 $applicationScope
+    ): void
+    {
+        $bitrix24Account = $this->createBitrix24AccountImplementation($uuid, $bitrix24UserId, $isBitrix24UserAdmin, $memberId, $domainUrl, $bitrix24AccountStatus, $authToken, $createdAt, $updatedAt, $applicationVersion, $applicationScope);
+        $bitrix24AccountRepository = $this->createBitrix24AccountRepositoryImplementation();
+
+        // application installed
+        $applicationToken = 'application_token';
+        $bitrix24Account->applicationInstalled($applicationToken);
+        $bitrix24AccountRepository->save($bitrix24Account);
+
+        // a few moments later
+        $account = $bitrix24AccountRepository->getById($uuid);
+        $account->applicationUninstalled($applicationToken);
+
+        $bitrix24AccountRepository->save($bitrix24Account);
+
+        $bitrix24AccountRepository->delete($uuid);
+
+        $this->expectException(Bitrix24AccountNotFoundException::class);
+        $bitrix24AccountRepository->getById($uuid);
+    }
+
+    /**
+     * @throws Bitrix24AccountNotFoundException
+     */
+    #[Test]
+    #[DataProvider('bitrix24AccountForInstallDataProvider')]
+    #[TestDox('test delete method for account not in deleted state')]
+    final public function testDeleteNotInDeletedState(
+        Uuid                  $uuid,
+        int                   $bitrix24UserId,
+        bool                  $isBitrix24UserAdmin,
+        string                $memberId,
+        string                $domainUrl,
+        Bitrix24AccountStatus $bitrix24AccountStatus,
+        AuthToken             $authToken,
+        CarbonImmutable       $createdAt,
+        CarbonImmutable       $updatedAt,
+        int                   $applicationVersion,
+        Scope                 $applicationScope
+    ): void
+    {
+        $bitrix24Account = $this->createBitrix24AccountImplementation($uuid, $bitrix24UserId, $isBitrix24UserAdmin, $memberId, $domainUrl, $bitrix24AccountStatus, $authToken, $createdAt, $updatedAt, $applicationVersion, $applicationScope);
+
+        $bitrix24AccountRepository = $this->createBitrix24AccountRepositoryImplementation();
+
+        $bitrix24AccountRepository->save($bitrix24Account);
+
+        $acc = $bitrix24AccountRepository->getById($uuid);
+        $this->assertEquals($bitrix24Account, $acc);
+
+        $this->expectException(InvalidArgumentException::class);
+        $bitrix24AccountRepository->delete($uuid);
+    }
+
+    /**
+     * @throws InvalidArgumentException
+     */
+    #[Test]
+    #[TestDox('test delete method with non existing account')]
+    public function testDeleteWithIdNotExists(): void
+    {
+        $this->expectException(Bitrix24AccountNotFoundException::class);
+        $bitrix24AccountRepository = $this->createBitrix24AccountRepositoryImplementation();
+        $bitrix24AccountRepository->delete(Uuid::v7());
+    }
+
+    #[Test]
+    #[TestDox('test findOneAdminByMemberId method with empty member_id')]
+    public function testFindOneAdminByMemberIdWithEmptyArgs(): void
+    {
+        $bitrix24AccountRepository = $this->createBitrix24AccountRepositoryImplementation();
+        $this->expectException(InvalidArgumentException::class);
+        /** @phpstan-ignore-next-line */
+        $bitrix24AccountRepository->findOneAdminByMemberId('');
+    }
+
+    /**
+     * @throws Bitrix24AccountNotFoundException
+     */
+    #[Test]
+    #[DataProvider('bitrix24AccountForInstallDataProvider')]
+    #[TestDox('test findOneAdminByMemberId method with happy path')]
+    final public function testFindOneAdminByMemberIdWithHappyPath(
+        Uuid                  $uuid,
+        int                   $bitrix24UserId,
+        bool                  $isBitrix24UserAdmin,
+        string                $memberId,
+        string                $domainUrl,
+        Bitrix24AccountStatus $bitrix24AccountStatus,
+        AuthToken             $authToken,
+        CarbonImmutable       $createdAt,
+        CarbonImmutable       $updatedAt,
+        int                   $applicationVersion,
+        Scope                 $applicationScope
+    ): void
+    {
+        $bitrix24Account = $this->createBitrix24AccountImplementation($uuid, $bitrix24UserId, $isBitrix24UserAdmin, $memberId, $domainUrl, $bitrix24AccountStatus, $authToken, $createdAt, $updatedAt, $applicationVersion, $applicationScope);
+        $bitrix24AccountRepository = $this->createBitrix24AccountRepositoryImplementation();
+
+        $bitrix24AccountRepository->save($bitrix24Account);
+
+        $acc = $bitrix24AccountRepository->getById($uuid);
+        $this->assertEquals($bitrix24Account, $acc);
+
+
+        $found = $bitrix24AccountRepository->findOneAdminByMemberId($memberId);
+        $this->assertEquals($acc, $found);
+    }
+
+    #[Test]
+    #[DataProvider('bitrix24AccountForInstallDataProvider')]
+    #[TestDox('test findOneAdminByMemberId method with simple user')]
+    final public function testFindOneAdminByMemberIdWithSimpleUser(
+        Uuid                  $uuid,
+        int                   $bitrix24UserId,
+        bool                  $isBitrix24UserAdmin,
+        string                $memberId,
+        string                $domainUrl,
+        Bitrix24AccountStatus $bitrix24AccountStatus,
+        AuthToken             $authToken,
+        CarbonImmutable       $createdAt,
+        CarbonImmutable       $updatedAt,
+        int                   $applicationVersion,
+        Scope                 $applicationScope
+    ): void
+    {
+        $bitrix24Account = $this->createBitrix24AccountImplementation($uuid, $bitrix24UserId, false, $memberId, $domainUrl, $bitrix24AccountStatus, $authToken, $createdAt, $updatedAt, $applicationVersion, $applicationScope);
+        $bitrix24AccountRepository = $this->createBitrix24AccountRepositoryImplementation();
+
+        $bitrix24AccountRepository->save($bitrix24Account);
+
+        $acc = $bitrix24AccountRepository->getById($uuid);
+        $this->assertEquals($bitrix24Account, $acc);
+
+        $found = $bitrix24AccountRepository->findOneAdminByMemberId($memberId);
+        $this->assertNull($found);
+    }
+
+    #[Test]
+    #[DataProvider('bitrix24AccountForInstallDataProvider')]
+    #[TestDox('test findByMemberId method with happy path')]
+    final public function testFindByMemberIdWithHappyPath(
+        Uuid                  $uuid,
+        int                   $bitrix24UserId,
+        bool                  $isBitrix24UserAdmin,
+        string                $memberId,
+        string                $domainUrl,
+        Bitrix24AccountStatus $bitrix24AccountStatus,
+        AuthToken             $authToken,
+        CarbonImmutable       $createdAt,
+        CarbonImmutable       $updatedAt,
+        int                   $applicationVersion,
+        Scope                 $applicationScope
+    ): void
+    {
+        $bitrix24Account = $this->createBitrix24AccountImplementation($uuid, $bitrix24UserId, $isBitrix24UserAdmin, $memberId, $domainUrl, $bitrix24AccountStatus, $authToken, $createdAt, $updatedAt, $applicationVersion, $applicationScope);
+        $bitrix24AccountRepository = $this->createBitrix24AccountRepositoryImplementation();
+
+        $bitrix24AccountRepository->save($bitrix24Account);
+
+        $acc = $bitrix24AccountRepository->getById($uuid);
+        $this->assertEquals($bitrix24Account, $acc);
+
+        $found = $bitrix24AccountRepository->findByMemberId($memberId);
+        $this->assertEquals($bitrix24Account, $found[0]);
+    }
+
+    #[Test]
+    #[TestDox('test findByMemberId method with happy path - not found')]
+    final public function testFindByMemberIdWithHappyPathNotFound(): void
+    {
+        $bitrix24AccountRepository = $this->createBitrix24AccountRepositoryImplementation();
+        $found = $bitrix24AccountRepository->findByMemberId('member_id');
+        $this->assertEquals([], $found);
+    }
+
+    #[Test]
+    #[TestDox('test findByMemberId method with empty member id')]
+    final public function testFindByMemberIdWithEmptyMemberId(): void
+    {
+        $bitrix24AccountRepository = $this->createBitrix24AccountRepositoryImplementation();
+        $this->expectException(InvalidArgumentException::class);
+        /** @phpstan-ignore-next-line */
+        $bitrix24AccountRepository->findByMemberId('');
+    }
+
+    /**
+     * @throws Bitrix24AccountNotFoundException
      * @throws InvalidArgumentException
      */
     #[Test]
     #[DataProvider('bitrix24AccountForInstallDataProvider')]
-    #[TestDox('test applicationInstalled method')]
-    final public function testApplicationInstalled(
+    #[TestDox('test findByMemberId method with blocked account happy path')]
+    final public function testFindByMemberIdWithBlockedAccountHappyPath(
         Uuid                  $uuid,
         int                   $bitrix24UserId,
         bool                  $isBitrix24UserAdmin,
@@ -304,56 +303,30 @@ abstract class Bitrix24AccountInterfaceTest extends TestCase
         CarbonImmutable       $createdAt,
         CarbonImmutable       $updatedAt,
         int                   $applicationVersion,
-        Scope                 $applicationScope,
-        string                $applicationToken,
-        ?Throwable            $throwable
+        Scope                 $applicationScope
     ): void
     {
-        if ($throwable instanceof \Throwable) {
-            $this->expectException($throwable::class);
-        }
-
         $bitrix24Account = $this->createBitrix24AccountImplementation($uuid, $bitrix24UserId, $isBitrix24UserAdmin, $memberId, $domainUrl, $bitrix24AccountStatus, $authToken, $createdAt, $updatedAt, $applicationVersion, $applicationScope);
-        $bitrix24Account->applicationInstalled($applicationToken);
-        $this->assertTrue($bitrix24Account->isApplicationTokenValid($applicationToken));
-    }
+        $bitrix24AccountRepository = $this->createBitrix24AccountRepositoryImplementation();
 
-    #[Test]
-    #[DataProvider('bitrix24AccountForUninstallDataProvider')]
-    #[TestDox('test applicationUninstalled method')]
-    final public function testApplicationUninstalled(
-        Uuid                  $uuid,
-        int                   $bitrix24UserId,
-        bool                  $isBitrix24UserAdmin,
-        string                $memberId,
-        string                $domainUrl,
-        Bitrix24AccountStatus $bitrix24AccountStatus,
-        AuthToken             $authToken,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
-        int                   $applicationVersion,
-        Scope                 $applicationScope,
-        string                $applicationToken,
-        ?Throwable            $throwable
-    ): void
-    {
-        if ($throwable instanceof \Throwable) {
-            $this->expectException($throwable::class);
-        }
+        $bitrix24AccountRepository->save($bitrix24Account);
 
-        $bitrix24Account = $this->createBitrix24AccountImplementation($uuid, $bitrix24UserId, $isBitrix24UserAdmin, $memberId, $domainUrl, $bitrix24AccountStatus, $authToken, $createdAt, $updatedAt, $applicationVersion, $applicationScope);
-        $bitrix24Account->applicationInstalled($applicationToken);
-        $bitrix24Account->applicationUninstalled($applicationToken);
-        $this->assertEquals(Bitrix24AccountStatus::deleted, $bitrix24Account->getStatus());
+        $acc = $bitrix24AccountRepository->getById($uuid);
+        $this->assertEquals($bitrix24Account, $acc);
+        $acc->markAsBlocked('block by admin');
+        $bitrix24AccountRepository->save($acc);
+
+        $found = $bitrix24AccountRepository->findByMemberId($memberId, Bitrix24AccountStatus::blocked);
+        $this->assertEquals($acc, $found[0]);
     }
 
     /**
-     * @throws InvalidArgumentException
+     * @throws Bitrix24AccountNotFoundException
      */
     #[Test]
-    #[DataProvider('bitrix24AccountWithStatusNewDataProvider')]
-    #[TestDox('test isApplicationTokenValid method')]
-    final public function testIsApplicationTokenValid(
+    #[DataProvider('bitrix24AccountForInstallDataProvider')]
+    #[TestDox('test findByMemberId method with account status but account not found')]
+    final public function testFindByMemberIdWithAccountStatusAccountNotFound(
         Uuid                  $uuid,
         int                   $bitrix24UserId,
         bool                  $isBitrix24UserAdmin,
@@ -364,112 +337,29 @@ abstract class Bitrix24AccountInterfaceTest extends TestCase
         CarbonImmutable       $createdAt,
         CarbonImmutable       $updatedAt,
         int                   $applicationVersion,
-        Scope                 $applicationScope,
-        string                $applicationToken,
+        Scope                 $applicationScope
     ): void
     {
         $bitrix24Account = $this->createBitrix24AccountImplementation($uuid, $bitrix24UserId, $isBitrix24UserAdmin, $memberId, $domainUrl, $bitrix24AccountStatus, $authToken, $createdAt, $updatedAt, $applicationVersion, $applicationScope);
-        $this->assertFalse($bitrix24Account->isApplicationTokenValid($applicationToken));
-        $bitrix24Account->applicationInstalled($applicationToken);
-        $this->assertTrue($bitrix24Account->isApplicationTokenValid($applicationToken));
+        $bitrix24AccountRepository = $this->createBitrix24AccountRepositoryImplementation();
+
+        $bitrix24AccountRepository->save($bitrix24Account);
+
+        $acc = $bitrix24AccountRepository->getById($uuid);
+        $this->assertEquals($bitrix24Account, $acc);
+
+        $found = $bitrix24AccountRepository->findByMemberId($memberId, Bitrix24AccountStatus::blocked);
+        $this->assertEquals([], $found);
     }
 
     /**
-     * @throws InvalidArgumentException
-     */
-    #[Test]
-    #[DataProvider('bitrix24AccountWithStatusNewDataProvider')]
-    #[TestDox('test getCreatedAt method')]
-    final public function testGetCreatedAt(
-        Uuid                  $uuid,
-        int                   $bitrix24UserId,
-        bool                  $isBitrix24UserAdmin,
-        string                $memberId,
-        string                $domainUrl,
-        Bitrix24AccountStatus $bitrix24AccountStatus,
-        AuthToken             $authToken,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
-        int                   $applicationVersion,
-        Scope                 $applicationScope,
-        string                $applicationToken,
-    ): void
-    {
-        $bitrix24Account = $this->createBitrix24AccountImplementation($uuid, $bitrix24UserId, $isBitrix24UserAdmin, $memberId, $domainUrl, $bitrix24AccountStatus, $authToken, $createdAt, $updatedAt, $applicationVersion, $applicationScope);
-        $this->assertTrue($bitrix24Account->getCreatedAt()->equalTo($createdAt));
-        $bitrix24Account->applicationInstalled($applicationToken);
-        $this->assertTrue($bitrix24Account->getCreatedAt()->equalTo($createdAt));
-    }
-
-    /**
-     * @throws InvalidArgumentException
-     */
-    #[Test]
-    #[DataProvider('bitrix24AccountWithStatusNewDataProvider')]
-    #[TestDox('test getUpdatedAt method')]
-    final public function testGetUpdatedAt(
-        Uuid                  $uuid,
-        int                   $bitrix24UserId,
-        bool                  $isBitrix24UserAdmin,
-        string                $memberId,
-        string                $domainUrl,
-        Bitrix24AccountStatus $bitrix24AccountStatus,
-        AuthToken             $authToken,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
-        int                   $applicationVersion,
-        Scope                 $applicationScope,
-        string                $applicationToken,
-    ): void
-    {
-        $bitrix24Account = $this->createBitrix24AccountImplementation($uuid, $bitrix24UserId, $isBitrix24UserAdmin, $memberId, $domainUrl, $bitrix24AccountStatus, $authToken, $createdAt, $updatedAt, $applicationVersion, $applicationScope);
-        $this->assertTrue($bitrix24Account->getUpdatedAt()->equalTo($updatedAt));
-        $bitrix24Account->applicationInstalled($applicationToken);
-        $this->assertFalse($bitrix24Account->getUpdatedAt()->equalTo($createdAt));
-    }
-
-    /**
-     * @throws InvalidArgumentException
-     */
-    #[Test]
-    #[DataProvider('bitrix24AccountWithStatusNewDataProvider')]
-    #[TestDox('test updateApplicationVersion method')]
-    final public function testUpdateApplicationVersion(
-        Uuid                  $uuid,
-        int                   $bitrix24UserId,
-        bool                  $isBitrix24UserAdmin,
-        string                $memberId,
-        string                $domainUrl,
-        Bitrix24AccountStatus $bitrix24AccountStatus,
-        AuthToken             $authToken,
-        CarbonImmutable       $createdAt,
-        CarbonImmutable       $updatedAt,
-        int                   $applicationVersion,
-        Scope                 $applicationScope,
-        string                $applicationToken,
-        int                   $newApplicationVersion,
-        Scope                 $newApplicationScope,
-        ?Throwable            $throwable
-    ): void
-    {
-        if ($throwable instanceof \Throwable) {
-            $this->expectException($throwable::class);
-        }
-
-        $bitrix24Account = $this->createBitrix24AccountImplementation($uuid, $bitrix24UserId, $isBitrix24UserAdmin, $memberId, $domainUrl, $bitrix24AccountStatus, $authToken, $createdAt, $updatedAt, $applicationVersion, $applicationScope);
-        $bitrix24Account->applicationInstalled($applicationToken);
-        $bitrix24Account->updateApplicationVersion($newApplicationVersion, $newApplicationScope);
-        $this->assertEquals($newApplicationVersion, $bitrix24Account->getApplicationVersion());
-        $this->assertTrue($newApplicationScope->equal($bitrix24Account->getApplicationScope()));
-    }
-
-    /**
+     * @throws Bitrix24AccountNotFoundException
      * @throws InvalidArgumentException
      */
     #[Test]
     #[DataProvider('bitrix24AccountForInstallDataProvider')]
-    #[TestDox('test markAsBlocked and getComment methods')]
-    final public function testMarkAsBlocked(
+    #[TestDox('test findByMemberId method with is admin happy path')]
+    final public function testFindByMemberIdWithIsAdminHappyPath(
         Uuid                  $uuid,
         int                   $bitrix24UserId,
         bool                  $isBitrix24UserAdmin,
@@ -480,31 +370,144 @@ abstract class Bitrix24AccountInterfaceTest extends TestCase
         CarbonImmutable       $createdAt,
         CarbonImmutable       $updatedAt,
         int                   $applicationVersion,
-        Scope                 $applicationScope,
-        string                $applicationToken,
-        ?Throwable            $throwable
+        Scope                 $applicationScope
     ): void
     {
-        if ($throwable instanceof \Throwable) {
-            $this->expectException($throwable::class);
-        }
-
         $bitrix24Account = $this->createBitrix24AccountImplementation($uuid, $bitrix24UserId, $isBitrix24UserAdmin, $memberId, $domainUrl, $bitrix24AccountStatus, $authToken, $createdAt, $updatedAt, $applicationVersion, $applicationScope);
-        $bitrix24Account->applicationInstalled($applicationToken);
+        $bitrix24AccountRepository = $this->createBitrix24AccountRepositoryImplementation();
 
-        $comment = 'block account just for fun';
-        $bitrix24Account->markAsBlocked($comment);
-        $this->assertEquals(Bitrix24AccountStatus::blocked, $bitrix24Account->getStatus());
-        $this->assertEquals($comment, $bitrix24Account->getComment());
+        $bitrix24AccountRepository->save($bitrix24Account);
+
+        $acc = $bitrix24AccountRepository->getById($uuid);
+        $this->assertEquals($bitrix24Account, $acc);
+
+        $found = $bitrix24AccountRepository->findByMemberId($memberId, null, true);
+        $this->assertEquals($acc, $found[0]);
     }
 
     /**
+     * @throws Bitrix24AccountNotFoundException
+     */
+    #[Test]
+    #[DataProvider('bitrix24AccountForInstallDataProvider')]
+    #[TestDox('test findByMemberId method with is admin - not found')]
+    final public function testFindByMemberIdWithIsAdminNotFound(
+        Uuid                  $uuid,
+        int                   $bitrix24UserId,
+        bool                  $isBitrix24UserAdmin,
+        string                $memberId,
+        string                $domainUrl,
+        Bitrix24AccountStatus $bitrix24AccountStatus,
+        AuthToken             $authToken,
+        CarbonImmutable       $createdAt,
+        CarbonImmutable       $updatedAt,
+        int                   $applicationVersion,
+        Scope                 $applicationScope
+    ): void
+    {
+        $bitrix24Account = $this->createBitrix24AccountImplementation($uuid, $bitrix24UserId, false, $memberId, $domainUrl, $bitrix24AccountStatus, $authToken, $createdAt, $updatedAt, $applicationVersion, $applicationScope);
+        $bitrix24AccountRepository = $this->createBitrix24AccountRepositoryImplementation();
+
+        $bitrix24AccountRepository->save($bitrix24Account);
+
+        $acc = $bitrix24AccountRepository->getById($uuid);
+        $this->assertEquals($bitrix24Account, $acc);
+
+        $found = $bitrix24AccountRepository->findByMemberId($memberId, null, true);
+        $this->assertEquals([], $found);
+    }
+
+    /**
+     * @throws Bitrix24AccountNotFoundException
+     */
+    #[Test]
+    #[DataProvider('bitrix24AccountForInstallDataProvider')]
+    #[TestDox('test findByMemberId method with all args')]
+    final public function testFindByMemberIdWithAllArgs(
+        Uuid                  $uuid,
+        int                   $bitrix24UserId,
+        bool                  $isBitrix24UserAdmin,
+        string                $memberId,
+        string                $domainUrl,
+        Bitrix24AccountStatus $bitrix24AccountStatus,
+        AuthToken             $authToken,
+        CarbonImmutable       $createdAt,
+        CarbonImmutable       $updatedAt,
+        int                   $applicationVersion,
+        Scope                 $applicationScope
+    ): void
+    {
+        $bitrix24Account = $this->createBitrix24AccountImplementation($uuid, $bitrix24UserId, false, $memberId, $domainUrl, $bitrix24AccountStatus, $authToken, $createdAt, $updatedAt, $applicationVersion, $applicationScope);
+        $bitrix24AccountRepository = $this->createBitrix24AccountRepositoryImplementation();
+
+        $bitrix24AccountRepository->save($bitrix24Account);
+
+        $acc = $bitrix24AccountRepository->getById($uuid);
+        $this->assertEquals($bitrix24Account, $acc);
+
+        $found = $bitrix24AccountRepository->findByMemberId($memberId, Bitrix24AccountStatus::new, false);
+        $this->assertEquals($acc, $found[0]);
+    }
+
+    /**
+     * @throws Bitrix24AccountNotFoundException
+     */
+    #[Test]
+    #[DataProvider('bitrix24AccountForInstallDataProvider')]
+    #[TestDox('test findByDomain method with happy path')]
+    final public function testFindByDomainWithHappyPath(
+        Uuid                  $uuid,
+        int                   $bitrix24UserId,
+        bool                  $isBitrix24UserAdmin,
+        string                $memberId,
+        string                $domainUrl,
+        Bitrix24AccountStatus $bitrix24AccountStatus,
+        AuthToken             $authToken,
+        CarbonImmutable       $createdAt,
+        CarbonImmutable       $updatedAt,
+        int                   $applicationVersion,
+        Scope                 $applicationScope
+    ): void
+    {
+        $bitrix24Account = $this->createBitrix24AccountImplementation($uuid, $bitrix24UserId, $isBitrix24UserAdmin, $memberId, $domainUrl, $bitrix24AccountStatus, $authToken, $createdAt, $updatedAt, $applicationVersion, $applicationScope);
+        $bitrix24AccountRepository = $this->createBitrix24AccountRepositoryImplementation();
+
+        $bitrix24AccountRepository->save($bitrix24Account);
+
+        $acc = $bitrix24AccountRepository->getById($uuid);
+        $this->assertEquals($bitrix24Account, $acc);
+
+        $found = $bitrix24AccountRepository->findByDomain($domainUrl);
+        $this->assertEquals($bitrix24Account, $found[0]);
+    }
+
+    #[Test]
+    #[TestDox('test findByDomain method with happy path - not found')]
+    final public function testFindByDomainWithHappyPathNotFound(): void
+    {
+        $bitrix24AccountRepository = $this->createBitrix24AccountRepositoryImplementation();
+        $found = $bitrix24AccountRepository->findByDomain('test.com');
+        $this->assertEquals([], $found);
+    }
+
+    #[Test]
+    #[TestDox('test findByDomain method with empty domain url')]
+    final public function testFindByDomainWithEmptyDomainUrl(): void
+    {
+        $bitrix24AccountRepository = $this->createBitrix24AccountRepositoryImplementation();
+        $this->expectException(InvalidArgumentException::class);
+        /** @phpstan-ignore-next-line */
+        $bitrix24AccountRepository->findByDomain('');
+    }
+
+    /**
+     * @throws Bitrix24AccountNotFoundException
      * @throws InvalidArgumentException
      */
     #[Test]
     #[DataProvider('bitrix24AccountForInstallDataProvider')]
-    #[TestDox('test markAsActive and getComment methods')]
-    final public function testMarkAsActive(
+    #[TestDox('test findByDomain method with blocked account happy path')]
+    final public function testFindByDomainWithBlockedAccountHappyPath(
         Uuid                  $uuid,
         int                   $bitrix24UserId,
         bool                  $isBitrix24UserAdmin,
@@ -515,26 +518,150 @@ abstract class Bitrix24AccountInterfaceTest extends TestCase
         CarbonImmutable       $createdAt,
         CarbonImmutable       $updatedAt,
         int                   $applicationVersion,
-        Scope                 $applicationScope,
-        string                $applicationToken,
-        ?Throwable            $throwable
+        Scope                 $applicationScope
     ): void
     {
-        if ($throwable instanceof \Throwable) {
-            $this->expectException($throwable::class);
-        }
-
         $bitrix24Account = $this->createBitrix24AccountImplementation($uuid, $bitrix24UserId, $isBitrix24UserAdmin, $memberId, $domainUrl, $bitrix24AccountStatus, $authToken, $createdAt, $updatedAt, $applicationVersion, $applicationScope);
-        $bitrix24Account->applicationInstalled($applicationToken);
+        $bitrix24AccountRepository = $this->createBitrix24AccountRepositoryImplementation();
 
-        $comment = 'block account just for fun';
-        $bitrix24Account->markAsBlocked($comment);
-        $this->assertEquals(Bitrix24AccountStatus::blocked, $bitrix24Account->getStatus());
-        $this->assertEquals($comment, $bitrix24Account->getComment());
-        $comment = 'activate account';
-        $bitrix24Account->markAsActive($comment);
-        $this->assertEquals(Bitrix24AccountStatus::active, $bitrix24Account->getStatus());
-        $this->assertEquals($comment, $bitrix24Account->getComment());
+        $bitrix24AccountRepository->save($bitrix24Account);
+
+        $acc = $bitrix24AccountRepository->getById($uuid);
+        $this->assertEquals($bitrix24Account, $acc);
+        $acc->markAsBlocked('block by admin');
+        $bitrix24AccountRepository->save($acc);
+
+        $found = $bitrix24AccountRepository->findByDomain($domainUrl, Bitrix24AccountStatus::blocked);
+        $this->assertEquals($acc, $found[0]);
+    }
+
+    /**
+     * @throws Bitrix24AccountNotFoundException
+     */
+    #[Test]
+    #[DataProvider('bitrix24AccountForInstallDataProvider')]
+    #[TestDox('test findByDomain method with account status but account not found')]
+    final public function testFindByDomainWithAccountStatusAccountNotFound(
+        Uuid                  $uuid,
+        int                   $bitrix24UserId,
+        bool                  $isBitrix24UserAdmin,
+        string                $memberId,
+        string                $domainUrl,
+        Bitrix24AccountStatus $bitrix24AccountStatus,
+        AuthToken             $authToken,
+        CarbonImmutable       $createdAt,
+        CarbonImmutable       $updatedAt,
+        int                   $applicationVersion,
+        Scope                 $applicationScope
+    ): void
+    {
+        $bitrix24Account = $this->createBitrix24AccountImplementation($uuid, $bitrix24UserId, $isBitrix24UserAdmin, $memberId, $domainUrl, $bitrix24AccountStatus, $authToken, $createdAt, $updatedAt, $applicationVersion, $applicationScope);
+        $bitrix24AccountRepository = $this->createBitrix24AccountRepositoryImplementation();
+
+        $bitrix24AccountRepository->save($bitrix24Account);
+
+        $acc = $bitrix24AccountRepository->getById($uuid);
+        $this->assertEquals($bitrix24Account, $acc);
+
+        $found = $bitrix24AccountRepository->findByDomain($domainUrl, Bitrix24AccountStatus::blocked);
+        $this->assertEquals([], $found);
+    }
+
+    /**
+     * @throws Bitrix24AccountNotFoundException
+     * @throws InvalidArgumentException
+     */
+    #[Test]
+    #[DataProvider('bitrix24AccountForInstallDataProvider')]
+    #[TestDox('test findByDomain method with is admin happy path')]
+    final public function testFindByDomainWithIsAdminHappyPath(
+        Uuid                  $uuid,
+        int                   $bitrix24UserId,
+        bool                  $isBitrix24UserAdmin,
+        string                $memberId,
+        string                $domainUrl,
+        Bitrix24AccountStatus $bitrix24AccountStatus,
+        AuthToken             $authToken,
+        CarbonImmutable       $createdAt,
+        CarbonImmutable       $updatedAt,
+        int                   $applicationVersion,
+        Scope                 $applicationScope
+    ): void
+    {
+        $bitrix24Account = $this->createBitrix24AccountImplementation($uuid, $bitrix24UserId, $isBitrix24UserAdmin, $memberId, $domainUrl, $bitrix24AccountStatus, $authToken, $createdAt, $updatedAt, $applicationVersion, $applicationScope);
+        $bitrix24AccountRepository = $this->createBitrix24AccountRepositoryImplementation();
+
+        $bitrix24AccountRepository->save($bitrix24Account);
+
+        $acc = $bitrix24AccountRepository->getById($uuid);
+        $this->assertEquals($bitrix24Account, $acc);
+
+        $found = $bitrix24AccountRepository->findByDomain($domainUrl, null, true);
+        $this->assertEquals($acc, $found[0]);
+    }
+
+    /**
+     * @throws Bitrix24AccountNotFoundException
+     */
+    #[Test]
+    #[DataProvider('bitrix24AccountForInstallDataProvider')]
+    #[TestDox('test findByDomain method with is admin - not found')]
+    final public function testFindByDomainWithIsAdminNotFound(
+        Uuid                  $uuid,
+        int                   $bitrix24UserId,
+        bool                  $isBitrix24UserAdmin,
+        string                $memberId,
+        string                $domainUrl,
+        Bitrix24AccountStatus $bitrix24AccountStatus,
+        AuthToken             $authToken,
+        CarbonImmutable       $createdAt,
+        CarbonImmutable       $updatedAt,
+        int                   $applicationVersion,
+        Scope                 $applicationScope
+    ): void
+    {
+        $bitrix24Account = $this->createBitrix24AccountImplementation($uuid, $bitrix24UserId, false, $memberId, $domainUrl, $bitrix24AccountStatus, $authToken, $createdAt, $updatedAt, $applicationVersion, $applicationScope);
+        $bitrix24AccountRepository = $this->createBitrix24AccountRepositoryImplementation();
+
+        $bitrix24AccountRepository->save($bitrix24Account);
+
+        $acc = $bitrix24AccountRepository->getById($uuid);
+        $this->assertEquals($bitrix24Account, $acc);
+
+        $found = $bitrix24AccountRepository->findByDomain($memberId, null, true);
+        $this->assertEquals([], $found);
+    }
+
+    /**
+     * @throws Bitrix24AccountNotFoundException
+     */
+    #[Test]
+    #[DataProvider('bitrix24AccountForInstallDataProvider')]
+    #[TestDox('test findByDomain method with all args')]
+    final public function testFindByDomainWithAllArgs(
+        Uuid                  $uuid,
+        int                   $bitrix24UserId,
+        bool                  $isBitrix24UserAdmin,
+        string                $memberId,
+        string                $domainUrl,
+        Bitrix24AccountStatus $bitrix24AccountStatus,
+        AuthToken             $authToken,
+        CarbonImmutable       $createdAt,
+        CarbonImmutable       $updatedAt,
+        int                   $applicationVersion,
+        Scope                 $applicationScope
+    ): void
+    {
+        $bitrix24Account = $this->createBitrix24AccountImplementation($uuid, $bitrix24UserId, false, $memberId, $domainUrl, $bitrix24AccountStatus, $authToken, $createdAt, $updatedAt, $applicationVersion, $applicationScope);
+        $bitrix24AccountRepository = $this->createBitrix24AccountRepositoryImplementation();
+
+        $bitrix24AccountRepository->save($bitrix24Account);
+
+        $acc = $bitrix24AccountRepository->getById($uuid);
+        $this->assertEquals($bitrix24Account, $acc);
+
+        $found = $bitrix24AccountRepository->findByDomain($domainUrl, Bitrix24AccountStatus::new, false);
+        $this->assertEquals($acc, $found[0]);
     }
 
     /**
@@ -573,7 +700,7 @@ abstract class Bitrix24AccountInterfaceTest extends TestCase
             new Scope(['crm', 'task']),
             'application_token',
             2,
-            new Scope(['task','crm']),
+            new Scope(['task', 'crm']),
             null
         ];
         yield 'valid-downgrade-scope' => [
@@ -590,7 +717,7 @@ abstract class Bitrix24AccountInterfaceTest extends TestCase
             new Scope([]),
             'application_token',
             2,
-            new Scope(['task','crm']),
+            new Scope(['task', 'crm']),
             null
         ];
         yield 'invalid-version' => [
@@ -698,22 +825,6 @@ abstract class Bitrix24AccountInterfaceTest extends TestCase
      */
     public static function bitrix24AccountForInstallDataProvider(): Generator
     {
-        yield 'empty-application-token' => [
-            Uuid::v7(),
-            12345,
-            true,
-            'member123',
-            'https://example.com',
-            Bitrix24AccountStatus::new,
-            new AuthToken('access_token', 'refresh_token', 1609459200),
-            CarbonImmutable::now(),
-            CarbonImmutable::now(),
-            1,
-            new Scope(['crm', 'task']),
-            '',
-            new InvalidArgumentException()
-        ];
-
         yield 'account-status-new' => [
             Uuid::v7(),
             12345,
@@ -728,51 +839,6 @@ abstract class Bitrix24AccountInterfaceTest extends TestCase
             new Scope(['crm', 'task']),
             'application_token_value',
             null
-        ];
-        yield 'account-status-active' => [
-            Uuid::v7(),
-            12345,
-            true,
-            'member123',
-            'https://example.com',
-            Bitrix24AccountStatus::active,
-            new AuthToken('access_token', 'refresh_token', 1609459200),
-            CarbonImmutable::now(),
-            CarbonImmutable::now(),
-            1,
-            new Scope(['crm', 'task']),
-            'application_token_value',
-            new InvalidArgumentException()
-        ];
-        yield 'account-status-blocked' => [
-            Uuid::v7(),
-            12345,
-            true,
-            'member123',
-            'https://example.com',
-            Bitrix24AccountStatus::blocked,
-            new AuthToken('access_token', 'refresh_token', 1609459200),
-            CarbonImmutable::now(),
-            CarbonImmutable::now(),
-            1,
-            new Scope(['crm', 'task']),
-            'application_token_value',
-            new InvalidArgumentException()
-        ];
-        yield 'account-status-deleted' => [
-            Uuid::v7(),
-            12345,
-            true,
-            'member123',
-            'https://example.com',
-            Bitrix24AccountStatus::deleted,
-            new AuthToken('access_token', 'refresh_token', 1609459200),
-            CarbonImmutable::now(),
-            CarbonImmutable::now(),
-            1,
-            new Scope(['crm', 'task']),
-            'application_token_value',
-            new InvalidArgumentException()
         ];
     }
